python-manager/modules/universal-converter/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
Pillow==10.2.0
pypandoc==1.12
pdf2docx==0.5.6
python-docx==1.1.0
pydantic==2.5.3
